Rename test to avoid conflict
